v0.6.0

+++ Fuwari/ViewController.swift
@@ -9,7 +9,7 @@
 import Cocoa
 import Quartz
 
-class ViewController: NSViewController {
+class ViewController: NSViewController, NSWindowDelegate {
 
     private var windowControllers = [FloatWindow]()
     private var isCancelled = false
@@ -19,6 +19,7 @@ class ViewController: NSViewController {
         super.viewDidLoad()
 
         NotificationCenter.default.addObserver(self, selector: #selector(startCapture), name: Notification.Name(rawValue: Constants.Notification.capture), object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(NSWindowDelegate.windowDidResize(_:)), name: NSWindow.didResizeNotification, object: nil)
         
         oldApp = NSWorkspace.shared.frontmostApplication
         oldApp?.activate(options: .activateIgnoringOtherApps)
@@ -28,11 +29,13 @@ class ViewController: NSViewController {
             let currentScreen = NSScreen.screens.first { $0.frame.contains(NSEvent.mouseLocation) }
             guard let currentScaleFactor = currentScreen?.backingScaleFactor else { return }
             let mouseLocation = NSEvent.mouseLocation
-            let ciImage = CIImage(contentsOf: imageUrl)?.copy() as? CIImage
-            let context = CIContext(options: nil)
-            let cgImage = context.createCGImage(ciImage!, from: ciImage!.extent)
+            guard let ciImage = CIImage(contentsOf: imageUrl)?.copy() as? CIImage else { return }
             
-            self.createFloatWindow(rect: NSRect(x: Int(mouseLocation.x) - cgImage!.width / Int(2 * currentScaleFactor), y: Int(mouseLocation.y) - cgImage!.height / Int(2 * currentScaleFactor), width: Int(CGFloat(cgImage!.width) / currentScaleFactor), height: Int(CGFloat(cgImage!.height) / currentScaleFactor)), image: cgImage!)
+            let context = CIContext(options: nil)
+            
+            guard let cgImage = context.createCGImage(ciImage, from: ciImage.extent) else { return }
+            
+            self.createFloatWindow(rect: NSRect(x: Int(mouseLocation.x) - cgImage.width / Int(2 * currentScaleFactor), y: Int(mouseLocation.y) - cgImage.height / Int(2 * currentScaleFactor), width: Int(CGFloat(cgImage.width) / currentScaleFactor), height: Int(CGFloat(cgImage.height) / currentScaleFactor)), image: cgImage)
             try? FileManager.default.removeItem(at: imageUrl)
         }
     }
@@ -51,6 +54,10 @@ class ViewController: NSViewController {
     @objc private func startCapture() {
         ScreenshotManager.shared.startCapture()
     }
+    
+    func windowDidResize(_ notification: Notification) {
+        windowControllers.filter {$0 .isKeyWindow}.first?.windowDidResize(notification)
+    }
 }
 
 extension ViewController: FloatDelegate {
@@ -58,7 +65,7 @@ extension ViewController: FloatDelegate {
         if !isCancelled {
             if windowControllers.filter({ $0 === floatWindow }).first != nil {
                 floatWindow.fadeWindow(isIn: false) {
-                    guard let index = self.windowControllers.index(where: {$0 === floatWindow}) else { return }
+                    guard let index = self.windowControllers.firstIndex(where: {$0 === floatWindow}) else { return }
                     self.windowControllers.remove(at: index)
                     self.windowControllers.last?.makeKey()
                 }
@@ -68,7 +75,7 @@ extension ViewController: FloatDelegate {
 
         if windowControllers.count == 0 {
             oldApp?.activate(options: .activateIgnoringOtherApps)
-        }        
+        }
     }
 
     func save(floatWindow: FloatWindow, image: CGImage) {
